sha1: move this to private websocket API
Nothing else uses it, and nothing else *should* use it because SHA1 is insecure.
WebSockets have to use it by definition, unfortunately.  The implementation is
not very fast, but doesn't have to be for the use case of websocket keying.
