+++ src/supplemental/websocket/sha1.c
