+++ src/supplemental/websocket/sha1.h
@@ -1,5 +1,5 @@
 //
-// Copyright 2018 Staysail Systems, Inc. <info@staysail.tech>
+// Copyright 2025 Staysail Systems, Inc. <info@staysail.tech>
 // Copyright 2018 Capitar IT Group BV <info@capitar.com>
 //
 // This software is supplied under the terms of the MIT License, a
@@ -8,8 +8,8 @@
 // found online at https://opensource.org/licenses/MIT.
 //
 
-#ifndef NNG_SUPPLEMENTAL_SHA1_SHA1_H
-#define NNG_SUPPLEMENTAL_SHA1_SHA1_H
+#ifndef NNG_SUPPLEMENTAL_WEBSOCKET_SHA1_H
+#define NNG_SUPPLEMENTAL_WEBSOCKET_SHA1_H
 
 typedef struct {
 	uint32_t digest[5]; // resulting digest
@@ -23,4 +23,4 @@ extern void nni_sha1_update(nni_sha1_ctx *, const void *, size_t);
 extern void nni_sha1_final(nni_sha1_ctx *, uint8_t[20]);
 extern void nni_sha1(const void *, size_t, uint8_t[20]);
 
-#endif // NNG_SUPPLEMENTAL_SHA1_SHA1_H
+#endif // NNG_SUPPLEMENTAL_WEBSOCKET_SHA1_H
